feat(Popover): refactor + draggable functionality
